pyproject.toml
[project]
name = "ai-engineering-robin-sundman-nilsson-de24"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.4.2",
    "fastapi>=0.122.0",
    "google-genai>=1.51.0",
    "google-generativeai>=0.8.5",
    "ipykernel>=7.1.0",
    "lancedb>=0.25.3",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pydantic-ai>=1.24.0",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
]
